requirements.txt
asgiref==3.2.10
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
Django==3.0.8
django-rest-auth==0.9.5
django-rest-authtoken==1.2.4
djangorestframework==3.11.0
firebase-admin==4.3.0
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-firestore==1.8.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
httplib2==0.18.1
idna==2.10
Jinja2==2.11.2
lazy-property==0.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
numpy==1.19.1
pandas==1.1.0
paytm-pg==1.0.9
paytmchecksum==1.7.0
pg-utils==0.6.0
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
python-dateutil==2.8.1
python-dotenv==0.14.0
python-http-client==3.2.7
pytz==2020.1
requests==2.24.0
rsa==4.6
sendgrid==6.4.4
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==16.7.9
wincertstore==0.2
